Tweaked handling of processor controls
Converted control values to double instead of float
Implemented double click to return to default value
Fixed viewport handling and component sizing
Updated example accordingly
Improved null handling of ProcessorHarness

+++ Source/Processing/ProcessorExamples.h
@@ -24,7 +24,7 @@ public:
     void process (const dsp::ProcessContextReplacing<float>& context) override;
     void reset() override;
     String getControlName (const int index) override;
-    float getDefaultControlValue (const int index);
+    double getDefaultControlValue (const int index);
 
 private:
     void init();
